requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
pypdf==4.2.0
openpyxl==3.1.2
Pillow>=10.4.0
gunicorn==22.0.0
